// frontend/src/lib/api/model/auditResultEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Good Doggy CRM API
 * Contrato de la API del backend. Fuente para el cliente TypeScript (orval).
 * OpenAPI spec version: 0.1.0
 */

/**
 * * `SUCCESS` - SUCCESS
 * * `DENIED` - DENIED
 * * `FAILED` - FAILED
 */
export type AuditResultEnum = (typeof AuditResultEnum)[keyof typeof AuditResultEnum];

export const AuditResultEnum = {
  SUCCESS: "SUCCESS",
  DENIED: "DENIED",
  FAILED: "FAILED",
} as const;